[feature][client] Verify checksum when available
It just report error when checksum differs
