+++ client/client.go
@@ -3,11 +3,9 @@ package client
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"os"
-	"path"
 	"path/filepath"
 	"strings"
 
@@ -15,6 +13,7 @@ import (
 	"github.com/progrhyme/binq"
 	"github.com/progrhyme/binq/internal/erron"
 	"github.com/progrhyme/binq/internal/logs"
+	"github.com/progrhyme/binq/schema/item"
 )
 
 type Mode int
@@ -35,6 +34,7 @@ type Runner struct {
 	ServerURL  *url.URL
 	httpClient *http.Client
 	sourceURL  string
+	sourceItem *item.ItemRevision
 	tmpdir     string
 	download   string
 	extractDir string
@@ -100,54 +100,6 @@ func (r *Runner) Run() (err error) {
 	if err = r.locate(); err != nil {
 		return err
 	}
-	return nil
-}
-
-func (r *Runner) fetch() (err error) {
-	if r.sourceURL == "" {
-		return fmt.Errorf("Can't fetch because sourceURL is not set. Source: %s", r.Source)
-	}
-	req, err := newHttpGetRequest(r.sourceURL, map[string]string{})
-	if err != nil {
-		return err
-	}
-	r.Logger.Printf("GET %s", r.sourceURL)
-	res, _err := r.httpClient.Do(req)
-	if _err != nil {
-		return erron.Errorwf(_err, "Failed to execute HTTP request")
-	}
-	defer res.Body.Close()
-	if res.StatusCode != 200 {
-		return fmt.Errorf("HTTP response is not OK. Code: %d, URL: %s", res.StatusCode, r.Source)
-	}
-	r.tmpdir, _err = ioutil.TempDir(os.TempDir(), "binq.*")
-	if _err != nil {
-		return erron.Errorwf(_err, "Failed to create tempdir")
-	}
-	defer func() {
-		if _err != nil {
-			os.RemoveAll(r.tmpdir)
-		}
-	}()
-
-	url, _err := url.Parse(r.sourceURL)
-	if _err != nil {
-		// Unexpected case
-		return erron.Errorwf(_err, "Failed to parse source URL: %v", r.sourceURL)
-	}
-	base := path.Base(url.Path)
-	r.download = filepath.Join(r.tmpdir, base)
-	dl, _err := os.Create(r.download)
-	if _err != nil {
-		return erron.Errorwf(_err, "Failed to open file: %s", r.download)
-	}
-	defer dl.Close()
-	_, _err = io.Copy(dl, res.Body)
-	if _err != nil {
-		return erron.Errorwf(_err, "Failed to read HTTP response")
-	}
-	r.Logger.Debugf("Saved file %s", r.download)
-
 	return nil
 }
